package update, lints, flame changes

+++ lib/src/extensions/point.dart
@@ -4,5 +4,5 @@ import "package:tmx_parser/tmx_parser.dart";
 /// Extension functions on [Point]
 extension PointExtensions on Point {
   /// Converts a Point to Vector2
-  Vector2 toVector2() => Vector2(this.x.toDouble(), this.y.toDouble());
+  Vector2 toVector2() => Vector2(this.x, this.y);
 }
